Closes #102
