+++ src/tools/EditableShape.js
@@ -1,4 +1,5 @@
 import EventEmitter from 'tiny-emitter';
+import { SVG_NAMESPACE } from '../util/SVG';
 
 const IMPLEMENTATION_MISSING = "An implementation is missing";
 
@@ -43,6 +44,50 @@ export default class EditableShape extends EventEmitter {
       
       this.resizeObserver.observe(this.svg.parentNode);
     }
+  }
+
+  drawHandle = (x, y) => {
+    const containerGroup = document.createElementNS(SVG_NAMESPACE, 'g');
+    containerGroup.setAttribute('class', 'a9s-handle');
+    containerGroup.setAttribute('transform-origin', `${x}px ${y}px`);
+
+    const group = document.createElementNS(SVG_NAMESPACE, 'g');
+    group.setAttribute('transform-origin', `${x}px ${y}px`);
+
+    const drawCircle = r => {
+      const c = document.createElementNS(SVG_NAMESPACE, 'circle');
+      c.setAttribute('cx', x);
+      c.setAttribute('cy', y);
+      c.setAttribute('r', r);
+      return c;
+    }
+
+    const radius = this.config.handleRadius || 6;
+
+    const inner = drawCircle(radius);
+    inner.setAttribute('class', 'a9s-handle-inner')
+
+    const outer = drawCircle(radius + 1);
+    outer.setAttribute('class', 'a9s-handle-outer')
+
+    group.appendChild(outer);
+    group.appendChild(inner);
+
+    containerGroup.appendChild(group);
+    return containerGroup;
+  }
+
+  setHandleXY = (handle, x, y) => {
+    handle.setAttribute('transform-origin', `${x}px ${y}px`);	
+    handle.firstChild.setAttribute('transform-origin', `${x}px ${y}px`);	
+
+    const inner = handle.querySelector('.a9s-handle-inner');	
+    inner.setAttribute('cx', x);	
+    inner.setAttribute('cy', y);	
+
+    const outer = handle.querySelector('.a9s-handle-outer');	
+    outer.setAttribute('cx', x);	
+    outer.setAttribute('cy', y);
   }
 
   scaleHandles = (scaleOrScaleX, optScaleY) => {
